fastapi/requirements.txt
fastapi
uvicorn
pymongo
google-generativeai
python-dotenv
pydantic
certifi
aiofiles
requests
gunicorn
